Discord Rich Presence support (#65)
* Initial Discord Rich Presence support

This adds installation and support for daglaroglou's league-rpc-client (https://github.com/daglaroglou/league-rpc-linux/), for Discord Rich Presence.
I didn't get daglaroglou's permission to add it, and there are some known issues (some distros don't use python3, others have a externally managed environment), but it should mostly work.

* fixing launch-script.py

it had a debug print, and a extra Popen

* Use venv for pip and python script

now externally managed packages should not be an issue anymore, since the install script creates a venv folder for the dependencies

* Make the RPC script installation happen in runtime

Now it checks if the folder is already installed, and if not it unchecks the box. If you check and apply, it will install the script for you.